Moved copy to separate folder and changed user repo to singular.
