+++ app/src/main/java/main/WTLibraryApp/Book/Copy/Copy.java
@@ -1,4 +1,4 @@
-package main.WTLibraryApp.Book;
+package main.WTLibraryApp.Book.Copy;
 
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
